Collect metadata spreadsheet code in a submodule
